Revert "Fix to use Int32"
This reverts commit bf75cc483a7bf23d16a79d340b0cacc4a7fb87bd.

+++ Source/Server/Modules/S_General.vb
@@ -58,7 +58,7 @@ Module S_General
 
         Console.WriteLine("Loading Character List...")
 
-        Dim ids As Task(Of List(Of Integer)) = GetData("account")
+        Dim ids As Task(Of List(Of Int64)) = GetData("account")
         Dim data As JObject
         Dim player As New PlayerStruct()
         Chars = New CharList
